+++ libraries/AP_Motors/AP_MotorsHeli_Single.h
@@ -61,7 +61,7 @@ public:
     void set_desired_rotor_speed(float desired_speed) override;
 
     // set_rpm - for rotor speed governor
-    void set_rpm(int16_t rotor_rpm) override;
+    void set_rpm(float rotor_rpm) override;
 
     // get_main_rotor_speed - estimated rotor speed when no speed sensor or governor is used
     float get_main_rotor_speed() const  override { return _main_rotor.get_rotor_speed(); }
